+++ src/pipt/update_schemes/esmda.py
@@ -169,7 +169,7 @@ class ESMDA(AssimilationScheme):
             self.trunc_energy = 0.98
 
         # Get the perturbed observations and observation scaling
-        self.vecObs = self.data_df.to_matrix()
+        self.vecObs = self.ensemble.obs_vector
         self.enObs = self.ensemble.perturb_observations(self.vecObs)
         self.enObs_conv = deepcopy(self.enObs)